+++ rust/telepathy-core/src/internal/utils.rs
@@ -24,7 +24,7 @@ use tokio::time::interval;
 use tokio_util::codec::LengthDelimitedCodec;
 use tokio_util::compat::FuturesAsyncReadCompatExt;
 use tokio_util::sync::CancellationToken;
-use tracing::debug;
+use tracing::{debug, error};
 #[cfg(target_family = "wasm")]
 use wasmtimer::tokio::interval;
 
@@ -218,4 +218,57 @@ pub(crate) fn select_best_connection(
             )
         })
         .map(|(id, s)| (*id, s))
+}
+
+#[cfg(target_os = "ios")]
+pub(crate) fn configure_audio_session() {
+    use objc2::runtime::{AnyObject, Bool};
+    use objc2::{class, msg_send};
+    use objc2_foundation::ns_string;
+
+    unsafe {
+        let av_audio_session: *mut AnyObject = msg_send![class!(AVAudioSession), sharedInstance];
+
+        // set category to `AVAudioSessionCategoryPlayAndRecord`
+        let category = ns_string!("AVAudioSessionCategoryPlayAndRecord");
+        let mode = ns_string!("AVAudioSessionModeDefault");
+        let error: *mut AnyObject = std::ptr::null_mut();
+
+        let success: Bool = msg_send![av_audio_session, setCategory: category,
+            mode: mode,
+            options: 0_u64,
+            error: &error];
+
+        if success == Bool::NO {
+            error!("Failed to set AVAudioSession category.");
+        }
+
+        let override_output: *mut AnyObject = msg_send![class!(AVAudioSession), sharedInstance];
+        let _: Bool = msg_send![override_output, overrideOutputAudioPort: 1_u64, error: &error];
+
+        // Activate the audio session
+        let success: Bool = msg_send![av_audio_session, setActive: Bool::YES, error: &error];
+
+        if success == Bool::NO {
+            error!("Failed to activate AVAudioSession.");
+        }
+    }
+}
+
+#[cfg(target_os = "ios")]
+pub(crate) fn deactivate_audio_session() {
+    use objc2::runtime::{AnyObject, Bool};
+    use objc2::{class, msg_send};
+    use objc2_foundation::ns_string;
+
+    unsafe {
+        let av_audio_session: *mut AnyObject = msg_send![class!(AVAudioSession), sharedInstance];
+
+        let error: *mut AnyObject = std::ptr::null_mut();
+        let success: Bool = msg_send![av_audio_session, setActive: Bool::NO, error: &error];
+
+        if success == Bool::NO {
+            error!("Failed to deactivate AVAudioSession.");
+        }
+    }
 }
